+++ front/src/DeckSelector/Components/LeftPanel.css
@@ -1,0 +1,54 @@
+:root {
+  --baseSpace: 2vw;
+}
+
+.body {
+  position: relative;
+  width: 50vw;
+  height: 50vh;
+  left: 25vw;
+  top: 25vh;
+  background-color: #936639;
+  color: white;
+  border-radius: 40px;
+  display: flex;
+  padding-top: var(--baseSpace);
+  padding-bottom: var(--baseSpace);
+  flex-direction: column;
+}
+
+.separator {
+  width: 100%;
+  height: 1px;
+  background-color: rgba(0, 0, 0, 0.2);
+}
+
+.background {
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.2);
+}
+
+.input {
+  width: 40vw;
+  background-color: #a68a64;
+  color: black;
+}
+
+.button:hover {
+  background-color: #414833;
+}
+
+.button {
+  background-color: #656d4a;
+  margin-top: 10px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  border-radius: 10px;
+}
+
+.button.click {
+  background-color: #333d29;
+}
