+++ front/src/JoinRoom/JoinRoom.jsx
@@ -1,68 +1,73 @@
-import React, { Component } from "react";
-import { Alert, Box, Button, Snackbar, Stack, Tab, Tabs, TextField } from "@mui/material";
+import React, { useState, useEffect } from "react";
+import {
+  Button,
+  Box,
+  Tabs,
+  Tab,
+  TextField,
+  Stack,
+  Snackbar,
+  Alert
+} from "@mui/material";
+import { useNavigate } from "react-router-dom";
 import { createRoomRequest, joinRoomRequest } from "../Api/roomRequest";
-import { withChannel } from "../Utility/hocChannel";
-import { withRouterProps } from "../Utility/hocNavigation";
-import { ROUTES } from "../Routes/routes";
-import { BaseLayout } from "../Components/Layouts/BaseLayout";
-import { TopBarButtonGroup } from "../Components/TopBar/TopBarButtonGroup";
-import { TopBarIconButton, TopBarTextButton } from "../Components/TopBar/TopBarButton";
-import { unselectGame } from "../Utility/navigate";
-import { Home } from "@mui/icons-material";
+import { JoinRoomNavigationBar } from "../NavigationBar/JoinRoomNavigationBar";
+import { useChannel } from "../ChannelContext";
 
-class JoinRoom extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      scenes: [],
-      selectedScene: "",
-      tabIndex: 0,
-      cards: [],
-      buttons: [],
-      roomId: "",
-      playerUsername: "",
-      snackbarOpen: false,
-      snackbarMessage: "",
-    };
-  }
+const JoinRoom = () => {
+  const navigate = useNavigate();
+  const { weakResetConnection } = useChannel();
+  const [scenes, setScenes] = useState([]);
+  const [selectedScene, setSelectedScene] = useState("");
+  const [tabIndex, setTabIndex] = useState(0);
+  const [cards, setCards] = useState([]);
+  const [buttons, setButtons] = useState([]);
+  const [roomId, setRoomId] = useState("");
+  const [playerUsername, setPlayerUsername] = useState("");
+  const [snackbarOpen, setSnackbarOpen] = useState(false);
+  const [snackbarMessage, setSnackbarMessage] = useState("");
+  const { setGameState } = useChannel()
 
-  componentDidMount() {
+  useEffect(() => {
     const savedScenes = JSON.parse(localStorage.getItem("scenes")) || [];
     if (savedScenes.length > 0) {
-      this.setState({ scenes: savedScenes, selectedScene: savedScenes[0] });
+      setScenes(savedScenes);
+      setSelectedScene(savedScenes[0]);
     }
-    document.title = "JCCE";
-  }
+  }, []);
 
-  componentDidUpdate(prevProps, prevState) {
-    const { selectedScene, cards, buttons, scenes } = this.state;
-
-    if (selectedScene !== prevState.selectedScene) {
+  useEffect(() => {
+    if (selectedScene) {
       const savedSceneData =
         JSON.parse(sessionStorage.getItem(selectedScene)) || { cards: [], buttons: [] };
-      this.setState({ cards: savedSceneData.cards, buttons: savedSceneData.buttons });
+      setCards(savedSceneData.cards);
+      setButtons(savedSceneData.buttons);
     }
+  }, [selectedScene]);
 
-    if (selectedScene && (cards !== prevState.cards || buttons !== prevState.buttons)) {
+  useEffect(() => {
+    if (selectedScene) {
       sessionStorage.setItem(selectedScene, JSON.stringify({ cards, buttons }));
     }
+  }, [cards, buttons, selectedScene]);
 
-    if (scenes !== prevState.scenes) {
+  useEffect(() => {
+    if (scenes.length > 0) {
       localStorage.setItem("scenes", JSON.stringify(scenes));
     }
-  }
+  }, [scenes]);
 
-  joinRoom = async () => {
-    const { weakResetConnection, setGameState } = this.props.channelContext;
-    const { roomId, playerUsername } = this.state;
+  useEffect(() => {
+    document.title = "JCCE";
+  }, []);
 
+  const joinRoom = async (navigate) => {
     try {
-      weakResetConnection();
-      const room_id = localStorage.getItem("room_id") ?? roomId;
+      weakResetConnection()
+      const room_id = localStorage.getItem("room_id") ?? roomId
       if (!room_id && !roomId.trim()) {
         throw new Error("Room ID is required");
       }
-
       let username = playerUsername.trim();
       if (!username) {
         const counter = parseInt(localStorage.getItem("playerCounter") || "1", 10);
@@ -72,164 +77,142 @@ class JoinRoom extends Component {
 
       const joinRoomFetch = async (username, roomId) => {
         try {
+          console.log(roomId, username)
           const response = await joinRoomRequest(roomId, { player_id: username });
-          setGameState(response);
+          setGameState(response)
+          console.log("response to join room:", response);
           return response;
         } catch (error) {
+          console.error("Error joining room:", error);
           throw error;
         }
       };
 
       try {
-        const data = await joinRoomFetch(username, room_id);
 
-        if (!data || !data.room_id) {
-          throw new Error("Failed to join room, no room_id returned");
-        }
+        const data = await joinRoomFetch(username, room_id);
         localStorage.setItem("room_id", room_id);
         localStorage.setItem("player_id", username);
         localStorage.setItem("playerUsername", username);
-        this.props.navigate(ROUTES.LOBBY);
+        navigate("/lobby");
       } catch (error) {
-        console.error("Error joining room:", error);
+        console.error("Join room failed with username", username, error);
         const currentCounter = parseInt(localStorage.getItem("playerCounter") || "1", 10);
         const newCounter = currentCounter + 1;
         const newUsername = `Player ${newCounter}`;
         localStorage.setItem("playerCounter", newCounter.toString());
         localStorage.setItem("playerUsername", newUsername);
         try {
-          console.log("Retrying with new username:", newUsername);
           const data = await joinRoomFetch(newUsername, room_id);
-          console.log("Joined room successfully with new username:", data);
           localStorage.setItem("player_id", data.player_id);
           localStorage.setItem("room_id", roomId);
-          this.props.navigate(ROUTES.LOBBY);
+          navigate("/lobby");
         } catch (err) {
+          console.error("Retry join room failed with username", newUsername, err);
           throw err;
         }
       }
     } catch (error) {
-      this.setState({ snackbarMessage: "Room ID not found", snackbarOpen: true });
+      console.error(error);
+      setSnackbarMessage("Room ID not found");
+      setSnackbarOpen(true);
     }
   };
 
-  createRoom = async () => {
+
+  const createRoom = async (navigate) => {
     try {
       const response = await createRoomRequest();
+      console.log("response to create room:", response);
       localStorage.setItem("room_id", response.room_id);
-      await this.joinRoom();
+      joinRoom(navigate)
     } catch (error) {
       console.error(error);
     }
   };
 
-  handleSnackbarClose = () => {
-    this.setState({ snackbarOpen: false });
+  const handleSnackbarClose = () => {
+    setSnackbarOpen(false);
   };
 
-  render() {
-    const { tabIndex, roomId, playerUsername, snackbarOpen, snackbarMessage } = this.state;
+  return (
+    <Box display="flex" flexDirection="column" height="100vh">
 
-    return (
-      <BaseLayout
-        topBar={
-          <TopBarButtonGroup>
-            <TopBarIconButton
-              event={() => {
-                unselectGame(this.props.navigate);
-              }}
-              svgComponent={Home}
-              altText="Return to home"
-            />
-            <TopBarTextButton
-              title={"Create/Join Room"}
-              altText={"Create or join a game room"}
-              event={() => this.props.navigate(ROUTES.JOIN)}
-            />
-            <TopBarTextButton
-              title={"Select Deck"}
-              altText={"Select a deck to play with"}
-              event={() => this.props.navigate(ROUTES.SELECT_DECK)}
-            />
-          </TopBarButtonGroup>
-        }
+      <JoinRoomNavigationBar />
 
-        centerPanel={
-          <>
-            <Box sx={styles.contentBox}>
-              <Tabs
-                value={tabIndex}
-                onChange={(_, newIndex) => this.setState({ tabIndex: newIndex })}
-                textColor="inherit"
-                indicatorColor="secondary"
+      <Box sx={styles.container}>
+        <Box sx={styles.contentBox}>
+          <Tabs
+            value={tabIndex}
+            onChange={(_, newIndex) => setTabIndex(newIndex)}
+            textColor="inherit"
+            indicatorColor="secondary"
+          >
+            <Tab label="Join Room" />
+            <Tab label="Create Room" />
+          </Tabs>
+          {tabIndex === 0 ? (
+            <Box sx={styles.formBox}>
+              <Stack spacing={2}>
+                <TextField
+                  fullWidth
+                  label="Enter Room ID"
+                  variant="outlined"
+                  sx={styles.textField}
+                  value={roomId}
+                  onChange={(e) => setRoomId(e.target.value)}
+                />
+                <TextField
+                  fullWidth
+                  label="Enter player username (optional)"
+                  variant="outlined"
+                  sx={styles.textField}
+                  value={playerUsername}
+                  onChange={(e) => setPlayerUsername(e.target.value)}
+                />
+              </Stack>
+              <Button
+                fullWidth
+                variant="contained"
+                sx={styles.button}
+                onClick={() => joinRoom(navigate)}
               >
-                <Tab label="Join Room"/>
-                <Tab label="Create Room"/>
-              </Tabs>
-              {tabIndex === 0 ? (
-                <Box sx={styles.formBox}>
-                  <Stack spacing={2}>
-                    <TextField
-                      fullWidth
-                      label="Enter Room ID"
-                      variant="outlined"
-                      sx={styles.textField}
-                      value={roomId}
-                      onChange={(e) => this.setState({ roomId: e.target.value })}
-                    />
-                    <TextField
-                      fullWidth
-                      label="Enter player username (optional)"
-                      variant="outlined"
-                      sx={styles.textField}
-                      value={playerUsername}
-                      onChange={(e) => this.setState({ playerUsername: e.target.value })}
-                    />
-                  </Stack>
-                  <Button
-                    fullWidth
-                    variant="contained"
-                    sx={styles.button}
-                    onClick={this.joinRoom}
-                  >
-                    Join Room
-                  </Button>
-                </Box>
-              ) : (
-                <Box sx={styles.formBox}>
-                  <TextField
-                    fullWidth
-                    label="Room Name"
-                    variant="outlined"
-                    sx={styles.textField}
-                  />
-                  <Button
-                    fullWidth
-                    variant="contained"
-                    sx={styles.button}
-                    onClick={this.createRoom}
-                  >
-                    Create Room
-                  </Button>
-                </Box>
-              )}
+                Join Room
+              </Button>
             </Box>
-            <Snackbar
-              open={snackbarOpen}
-              autoHideDuration={6000}
-              onClose={this.handleSnackbarClose}
-              anchorOrigin={{ vertical: "bottom", horizontal: "center" }}
-            >
-              <Alert onClose={this.handleSnackbarClose} severity="error" sx={{ width: "100%" }}>
-                {snackbarMessage}
-              </Alert>
-            </Snackbar>
-          </>
-        }
-      />
-    );
-  }
-}
+          ) : (
+            <Box sx={styles.formBox}>
+              <TextField
+                fullWidth
+                label="Room Name"
+                variant="outlined"
+                sx={styles.textField}
+              />
+              <Button
+                fullWidth
+                variant="contained"
+                sx={styles.button}
+                onClick={() => createRoom(navigate)}
+              >
+                Create Room
+              </Button>
+            </Box>
+          )}
+        </Box>
+      </Box>
+      <Snackbar
+        open={snackbarOpen}
+        autoHideDuration={6000}
+        onClose={handleSnackbarClose}
+        anchorOrigin={{ vertical: "bottom", horizontal: "center" }}
+      >
+        <Alert onClose={handleSnackbarClose} severity="error" sx={{ width: "100%" }}>
+          {snackbarMessage}
+        </Alert>
+      </Snackbar>
+    </Box>
+  );
+};
 
 const styles = {
   navbar: {
@@ -275,4 +258,4 @@ const styles = {
   },
 };
 
-export default withChannel(withRouterProps(JoinRoom));
+export default JoinRoom;
