+++ front/src/NavigationBar/RoomNavigationBar.jsx
@@ -1,9 +1,10 @@
 import React, { useState, useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { Home, Chat as ChatIcon } from '@mui/icons-material'
-import { NavbarButton, NavbarSmallButton } from './navbarButton'
+import { NavbarButton, NavbarSmallButton } from './Components/navbarButton'
 import { Box, Button, Popover, TextField, Typography } from '@mui/material'
 import LinkIcon from '@mui/icons-material/Link'
+import GameChat from '../Room/Componnent/GameChat'
 
 export const RoomNavigationBar = ({ roomId }) => {
   const navigate = useNavigate()
