+++ front/src/Routes/AppRoutes.jsx
@@ -1,5 +1,4 @@
 import { Route, Routes } from "react-router-dom";
-import GameList from "../GameList/GameList";
 import GameMainPage from "../GameMainPage/GameMainPage";
 import CardEditor from "../CardEditor/CardEditor";
 import TypeEditor from "../TypeEditor/TypeEditor";
@@ -14,6 +13,8 @@ import Lobby from "../Lobby/Lobby";
 import { HashRouter as Router } from "react-router";
 import React from "react";
 import { ROUTES } from "./routes";
+import Community from "../Community/Community";
+import Templates from "../Templates/Templates";
 
 export class AppRoutes extends React.Component {
 
@@ -21,18 +22,22 @@ export class AppRoutes extends React.Component {
     return (
       <Router>
         <Routes>
-          <Route path={ROUTES.HOME} element={<GameList/>}/>
+          <Route path={ROUTES.HOME} element={<Community/>}/>
+
           <Route path={ROUTES.GAME_MAIN_PAGE} element={<GameMainPage/>}/>
           <Route path={ROUTES.CARD_EDITOR} element={<CardEditor/>}/>
           <Route path={ROUTES.TYPE_EDITOR} element={<TypeEditor/>}/>
           <Route path={ROUTES.BOARD_EDITOR} element={<BoardEditor/>}/>
+          <Route path={ROUTES.TEMPLATES} element={<Templates/>}/>
+
           <Route path={ROUTES.JOIN} element={<JoinRoom/>}/>
-          <Route path={ROUTES.LOGIN} element={<Login/>}/>
-          <Route path={ROUTES.HELP_GAME} element={<HelpGame/>}/>
-          <Route path={ROUTES.ROOM} element={<Room/>}/>
           <Route path={ROUTES.SELECT_DECK} element={<DeckSelector/>}/>
           <Route path={ROUTES.EDIT_DECK} element={<DeckBuilder/>}/>
           <Route path={ROUTES.LOBBY} element={<Lobby/>}/>
+
+          <Route path={ROUTES.ROOM} element={<Room/>}/>
+          <Route path={ROUTES.LOGIN} element={<Login/>}/>
+          <Route path={ROUTES.HELP_GAME} element={<HelpGame/>}/>
         </Routes>
       </Router>
     )
