+++ front/src/Routes/routes.js
@@ -1,14 +1,18 @@
 export const ROUTES = {
   HOME: "/",
+
   GAME_MAIN_PAGE: "/game-main-page",
   CARD_EDITOR: "/card-editor",
   TYPE_EDITOR: "/type-editor",
   BOARD_EDITOR: "/board-editor",
+  TEMPLATES: "/templates",
+
   JOIN: "/join",
-  LOGIN: "/login",
-  HELP_GAME: "/help-game",
-  ROOM: "/room",
   SELECT_DECK: "/select-deck",
   EDIT_DECK: "/edit-deck",
   LOBBY: "/lobby",
+
+  ROOM: "/room",
+  LOGIN: "/login",
+  HELP_GAME: "/help-game",
 }
